+++ packages/xchain-thorchain-query/src/types.ts
@@ -143,7 +143,9 @@ export type EstimateAddLP = {
   slipPercent: BigNumber
   poolShare: PoolShareDetail
   lpUnits: BaseAmount
-  transactionFee: LPFees
+  inbound: {
+    fees: LPAmounts
+  }
   runeToAssetRatio: BigNumber
   estimatedWaitSeconds: number
   errors: string[]
@@ -153,19 +155,24 @@ export type EstimateWithdrawLP = {
   assetAddress?: string
   runeAddress?: string
   slipPercent: BigNumber
-  transactionFee: LPFees
+  inbound: {
+    minToSend: LPAmounts
+    fees: LPAmounts
+  }
+  outboundFee: LPAmounts
   assetAmount: CryptoAmount
   runeAmount: CryptoAmount
+  lpGrowth: string
   impermanentLossProtection: ILProtectionData
   estimatedWaitSeconds: number
   assetPool: string
 }
-
-export type LPFees = {
-  runeFee: CryptoAmount
-  assetFee: CryptoAmount
-  totalFees: CryptoAmount
+export type LPAmounts = {
+  rune: CryptoAmount
+  asset: CryptoAmount
+  total: CryptoAmount
 }
+
 export type DustValues = {
   asset: CryptoAmount
   rune: CryptoAmount
@@ -183,6 +190,7 @@ export type WithdrawLiquidityPosition = {
 export type LiquidityPosition = {
   poolShare: PoolShareDetail
   position: LiquidityProvider
+  lpGrowth: string
   impermanentLossProtection: ILProtectionData
 }